pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caulking"
version = "0.2.2"
description = "Enforce blocking secret scans (gitleaks-first) + ecosystem guardrails via pre-commit."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "CC0" }
authors = [{ name = "Cloud.gov Security", email = "security@cloud.gov" }]
keywords = [
    "security",
    "secrets",
    "pre-commit",
    "gitleaks",
    "detect-secrets",
    "guardrails",
]
dependencies = ["typer>=0.12.5", "PyYAML>=6.0.2"]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.1",
    "mypy>=1.11.2",
    "types-PyYAML>=6.0.12.20240808",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "bandit>=1.7.10",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["caulking*"]

[tool.ruff]
target-version = "py39"
line-length = 100
extend-exclude = ["build", "dist", ".venv"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
    "B",
    "A",
    "C90",
    "RUF",
    "PL",
    "SIM",
    "PT",
    "S",
]
ignore = ["PLR0912", "PLR0913"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
"src/caulking/cli.py" = ["C901"]

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
strict_equality = true
check_untyped_defs = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[tool.bandit]
targets = ["src"]
severity = "MEDIUM"
confidence = "HIGH"
skips = []

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
branch = true
source = ["src/caulking"]
omit = ["src/caulking/__main__.py"]

[tool.coverage.report]
show_missing = true
skip_covered = true
